Move system_memory_cq_interface
