+++ tt_metal/impl/dispatch/system_memory_manager.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 // needed for private members
-#include <tt-metalium/system_memory_cq_interface.hpp>
+#include "system_memory_cq_interface.hpp"
 #include <umd/device/chip_helpers/tlb_manager.h>  // needed because tt_io.hpp requires needs TLBManager
 #include <umd/device/tt_io.hpp>                   // for tt::Writer
 #include <umd/device/tt_xy_pair.h>                // for tt_cxy_pair
